refactor(match): 매칭 데이터 변환 로직의 역할 재정의 및 명칭 변경
