+++ src/main/java/com/likelion/zooting/domain/match/policy/MatchScoreCalculatePolicy.java
@@ -1,8 +1,8 @@
 package com.likelion.zooting.domain.match.policy;
 
-import com.likelion.zooting.domain.match.dto.data.UserAnimalMatchCandidate;
-import com.likelion.zooting.domain.match.dto.data.UserInterestMatchCandidate;
-import com.likelion.zooting.domain.match.dto.data.UserMovieGenreMatchCandidate;
+import com.likelion.zooting.domain.match.service.data.UserAnimalMatchCandidate;
+import com.likelion.zooting.domain.match.service.data.UserInterestMatchCandidate;
+import com.likelion.zooting.domain.match.service.data.UserMovieGenreMatchCandidate;
 
 import java.util.List;
 import java.util.Map;
