+++ src/test/java/com/likelion/zooting/domain/match/policy/impl/MatchScoreCalculatePolicyImplTest.java
@@ -1,8 +1,8 @@
 package com.likelion.zooting.domain.match.policy.impl;
 
-import com.likelion.zooting.domain.match.dto.data.UserAnimalMatchCandidate;
-import com.likelion.zooting.domain.match.dto.data.UserInterestMatchCandidate;
-import com.likelion.zooting.domain.match.dto.data.UserMovieGenreMatchCandidate;
+import com.likelion.zooting.domain.match.service.data.UserAnimalMatchCandidate;
+import com.likelion.zooting.domain.match.service.data.UserInterestMatchCandidate;
+import com.likelion.zooting.domain.match.service.data.UserMovieGenreMatchCandidate;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
